requirements.txt
matplotlib==3.0.0
numpy==1.15.2
opencv-python==3.4.3.18
requests==2.20.0
six==1.11.0
urllib3==1.23
flask-restful
simplejson
imutils

kafka-python

pillow
flask-sqlalchemy
flask-migrate
pymysql